refactor(server): share bucket artifact retention scanning
