+++ server/lib/tuist/storage/cache_artifact_retention.ex
@@ -1,150 +1,24 @@
 defmodule Tuist.Storage.CacheArtifactRetention do
   @moduledoc false
 
-  import Ecto.Query
-
-  alias Tuist.Accounts.Account
   alias Tuist.Environment
-  alias Tuist.Repo
-  alias Tuist.Storage
-  alias Tuist.Storage.RetentionPolicy
+  alias Tuist.Storage.BucketArtifactRetention
 
-  @default_page_size 1000
-  @delete_receive_timeout 60_000
-  @delete_task_timeout 65_000
   @artifact_types [:xcode_cache, :cas, :xcode_module, :gradle]
 
   def artifact_types, do: @artifact_types
 
   def delete_expired(artifact_type, opts \\ []) when artifact_type in @artifact_types do
-    target = retention_target(artifact_type)
-    bucket_name = Map.fetch!(target, :bucket_name)
-
-    if is_nil(bucket_name) or bucket_name == "" do
-      :ok
-    else
-      page_size = Keyword.get(opts, :page_size, @default_page_size)
-      continuation_token = Keyword.get(opts, :continuation_token)
-
-      case Storage.list_objects_from_bucket(bucket_name,
-             prefix: "",
-             max_keys: page_size,
-             continuation_token: continuation_token
-           ) do
-        {:ok, %{body: body}} ->
-          objects = Map.get(body, :contents, [])
-
-          objects_to_delete =
-            objects
-            |> Enum.filter(&matches_retention_target?(&1, target))
-            |> expired_objects(target)
-
-          delete_opts = [receive_timeout: @delete_receive_timeout, task_timeout: @delete_task_timeout]
-
-          with :ok <- Storage.delete_objects_from_bucket(Enum.map(objects_to_delete, & &1.key), bucket_name, delete_opts) do
-            {:ok, next_continuation_token(body)}
-          end
-
-        {:error, reason} ->
-          {:error, reason}
-      end
-    end
+    artifact_type
+    |> retention_target()
+    |> BucketArtifactRetention.delete_expired(opts)
   end
-
-  defp expired_objects(objects, target) do
-    plans_by_account_handle = managed_plans_by_account_handle(objects)
-    retention_artifact_type = Map.fetch!(target, :retention_artifact_type)
-
-    Enum.filter(objects, fn object ->
-      # Managed retention only deletes objects whose account resolves to the
-      # managed storage path. Unknown and custom-storage accounts are skipped.
-      case plan_for_object(object, plans_by_account_handle) do
-        nil ->
-          false
-
-        plan ->
-          cutoff = RetentionPolicy.cutoff(retention_artifact_type, plan)
-
-          object
-          |> last_modified()
-          |> expired?(cutoff)
-      end
-    end)
-  end
-
-  defp managed_plans_by_account_handle(objects) do
-    account_handles =
-      objects
-      |> Enum.map(&account_handle/1)
-      |> Enum.reject(&is_nil/1)
-      |> Enum.flat_map(fn account_handle -> [account_handle, normalize_account_handle(account_handle)] end)
-      |> Enum.uniq()
-
-    Account
-    |> where([account], account.name in ^account_handles)
-    |> Repo.all()
-    |> Enum.reject(&Account.custom_s3_storage_configured?/1)
-    |> then(fn accounts ->
-      # `current_plans/1` batch-loads subscriptions for all fetched accounts,
-      # keeping this at two queries per S3 page instead of one query per account.
-      plans_by_account_id = RetentionPolicy.current_plans(accounts)
-
-      exact_plans_by_account_handle =
-        Map.new(accounts, fn account -> {account.name, Map.fetch!(plans_by_account_id, account.id)} end)
-
-      normalized_plans_by_account_handle =
-        Map.new(accounts, fn account ->
-          {normalize_account_handle(account.name), Map.fetch!(plans_by_account_id, account.id)}
-        end)
-
-      Map.merge(normalized_plans_by_account_handle, exact_plans_by_account_handle)
-    end)
-  end
-
-  defp plan_for_object(object, plans_by_account_handle) do
-    account_handle = account_handle(object)
-
-    Map.get(plans_by_account_handle, account_handle) ||
-      Map.get(plans_by_account_handle, normalize_account_handle(account_handle))
-  end
-
-  defp matches_retention_target?(object, target) do
-    expected_path_segment = Map.fetch!(target, :object_path_segment)
-
-    case String.split(object.key, "/", parts: 4) do
-      [_account_handle, _project_handle, path_segment, _rest] -> path_segment == expected_path_segment
-      _ -> false
-    end
-  end
-
-  defp account_handle(%{key: key}) do
-    case String.split(key, "/", parts: 2) do
-      [account_handle, _rest] -> account_handle
-      _ -> nil
-    end
-  end
-
-  defp normalize_account_handle(nil), do: nil
-  defp normalize_account_handle(account_handle), do: String.downcase(account_handle)
-
-  defp last_modified(%{last_modified: %DateTime{} = last_modified}), do: last_modified
-
-  defp last_modified(%{last_modified: last_modified}) when is_binary(last_modified) do
-    case DateTime.from_iso8601(last_modified) do
-      {:ok, datetime, _offset} -> datetime
-      {:error, _reason} -> nil
-    end
-  end
-
-  defp last_modified(_object), do: nil
-
-  defp expired?(nil, _cutoff), do: false
-  defp expired?(last_modified, cutoff), do: DateTime.before?(last_modified, cutoff)
 
   defp retention_target(:xcode_cache) do
     %{
       bucket_name: Environment.cache_xcode_s3_bucket_name(),
-      object_path_segment: "xcode",
+      object_matches?: object_path_segment_matches?("xcode"),
+      skip_custom_storage_accounts?: true,
       retention_artifact_type: :xcode_cache_artifact
     }
   end
@@ -152,7 +26,8 @@ defmodule Tuist.Storage.CacheArtifactRetention do
   defp retention_target(:cas) do
     %{
       bucket_name: Environment.cache_s3_bucket_name(),
-      object_path_segment: "cas",
+      object_matches?: object_path_segment_matches?("cas"),
+      skip_custom_storage_accounts?: true,
       retention_artifact_type: :xcode_cache_artifact
     }
   end
@@ -160,7 +35,8 @@ defmodule Tuist.Storage.CacheArtifactRetention do
   defp retention_target(:xcode_module) do
     %{
       bucket_name: Environment.cache_s3_bucket_name(),
-      object_path_segment: "module",
+      object_matches?: object_path_segment_matches?("module"),
+      skip_custom_storage_accounts?: true,
       retention_artifact_type: :cache_artifact
     }
   end
@@ -168,14 +44,18 @@ defmodule Tuist.Storage.CacheArtifactRetention do
   defp retention_target(:gradle) do
     %{
       bucket_name: Environment.cache_s3_bucket_name(),
-      object_path_segment: "gradle",
+      object_matches?: object_path_segment_matches?("gradle"),
+      skip_custom_storage_accounts?: true,
       retention_artifact_type: :cache_artifact
     }
   end
 
-  defp next_continuation_token(body) do
-    if Map.get(body, :is_truncated) in [true, "true"] do
-      Map.get(body, :next_continuation_token)
+  defp object_path_segment_matches?(expected_path_segment) do
+    fn object ->
+      case String.split(object.key, "/", parts: 4) do
+        [_account_handle, _project_handle, path_segment, _rest] -> path_segment == expected_path_segment
+        _ -> false
+      end
     end
   end
 end
